finish module renaming
